fix: implement full contract editor functionality in create page
This commit resolves critical issues with the contract creation functionality by properly implementing the CreateContractPage component:

- Fix incomplete page implementation that was showing only a placeholder instead of the full editor interface
- Properly wrap useSearchParams hook in a Suspense boundary to prevent build errors
- Implement all required component functions for contract validation, compilation and deployment
- Fix React Hook dependency warnings by correctly managing dependencies in useCallback hooks
- Add proper toast notifications for user feedback during validation, compilation and deployment
- Ensure correct state management for all editor operations
- Implement proper error handling with user-friendly explanations
- Structure the codebase to follow React best practices and improve build process

These changes were necessary because the previous implementation had a placeholder comment that was never replaced with actual code, causing the contract editor to not render properly. The build was also failing due to improper use of the useSearchParams hook.

The new implementation ensures a complete, functioning smart contract editor with all the intended features including code editing, validation, compilation, and deployment.
